[TC] Regenerated demo for tc_tc_compare_mode_sam_a5d2_xult
MH3-67224

+++ apps/tc/tc_compare_mode/firmware/src/config/sam_a5d2_xult/peripheral/tc/plib_tc1.c
@@ -112,7 +112,7 @@ void TC1_CH0_ComparePeriodSet (uint32_t period)
 /* Read the period value */
 uint32_t TC1_CH0_ComparePeriodGet (void)
 {
-    return TC1_REGS->TC_CHANNEL[0].TC_RC;
+    return (uint32_t)TC1_REGS->TC_CHANNEL[0].TC_RC;
 }
 
 /* Set the compare A value */
